requirements.txt
pandas
pyyaml
jsons 
requests
lxml
owslib